Fix "all" calendar
The calendar defined by `all.toml` does not build correctly because
it requires two files that do not exist:

- `language.toml`
- `library.toml`

The corresponding files were instead named:

- `lang.toml`
- `libs.toml`

These names are better as they match the conventional names of the
teams.  So let's fix the "all" calendar by updating it to refer to
these files.