Rename `infra` more conventionally
The calendars for all the teams represented here follow the
conventional naming scheme for each team (e.g. as represented by the
name of that team's own organizational git repository) with one
exception.  We had named the calendar for the infrastructure team
`infrastructure.toml` rather than `infra.toml`.

The infrastructure team is conventionally known as `t-infra` or as the
`infra-team`.  E.g., the organizational repository for the team has
the URL:

    https://github.com/rust-lang/infra-team

With that in mind, let's fix this while the calendaring is still young
and rename `infrastructure.toml` to `infra.toml`.