Mark subscribers as busy during T-lang, WG-async events
In many kinds of calendaring software, a person working in an
organization will make his or her calendar available to others in such
a way that those others can try to schedule around times when that
person is busy.  To do this, calendaring software distinguishes
between events that make the person busy and those that don't.

In the iCalendar standard, this is termed the "transparency" of the
event.[^1]  An event which is "opaque" is one that consumes the
attention of that individual.

The tooling here[^2] does not emit any "transparency" property in the
output ICS file by default.  If we want this property emitted, we need
to specify it.  Without it specified, calendaring software may use
whatever its own default is.

The lang team meetings and those of WG-async are fully consuming.
Let's mark those as opaque.

[^1]: https://datatracker.ietf.org/doc/html/rfc5545#section-3.8.2.7

[^2]: https://github.com/rust-lang/calendar-generation/