Extend T-lang design/planning meetings to 90 minutes
We decided to extend our design meetings to 90 minutes and to start
each half an hour earlier.  To preserve consistency, let's do the same
for the planning meetings.  Those often will not need the full 90
minutes on most months, but given the upcoming edition we may in fact
need this, and we can of course hand the time back to people.

To implement this, we create two new events and we mark the existing
events as having ended.  We do it this way as modifying the existing
events would retroactively change the time of the events that have
already occurred.

For the new events, we use 160-bit random identifiers to ensure that
these are globally-unique as required by IETF RFC 5545.