Update WG-async meetings for time and TZ change, etc.
The WG-async meetings had been scheduled on UTC time on this
calendar.  However, that's probably not what we had intended, so let's
fix that.

We'll instead schedule these meetings on U.S. time (and thus follow
the schedule for U.S. DST) as that's most common for the project, and
in particular, because that's how the Monday T-types and the Thursday
T-spec meetings are scheduled, and we don't want to jump around with
respect to those.

For the Thursday meeting, however, we'd been discussing moving that
back an hour anyway.  The effect of DST pushed it back an hour in
March, and we'll keep with that when converting the timezone.  To
preserve the correctness of earlier calendar events, we'll retire the
old event and create a new one.  In doing this, we'll go ahead and
separate out the design and planning meetings into two different
calendar events, as that's how it's done on the lang calendar.

For the triage meeting, on the other hand, we continued meeting on
U.S. time in spite of what was on the calendar, so we can go ahead
and just update the existing event to make the time of those meetings
retroactively correct.

While we're doing updates, we'll go ahead and also update the
locations and descriptions to be more correct and to the point.