End broken lang planning meeting event
We have a lang planning meeting scheduled for the first Wednesday of
every month.  This was set up as a monthly-recurring event starting on
2024-03-06, which was a Wednesday.

However, what that actually encodes is a meeting every month on the
6th.  That's not what we want and has been confusing people, so let's
retroactively end that recurrence.

We don't need any replacement here, as we actually just use our design
meeting slot for this, and lately we've been mixing it up a bit and
not always doing these on the first Wednesday of the month.