Add xtask tool to generate event UIDs
According to the iCalendar (IETF) RFC 5545, the UID for each event
must be "globally unique".  In this context, that means that the
identifier must not collide with any other identifier for any other
calendar event across all calendars, ever.

The way we were suggesting that people generate these, by calling out
to `date`, unfortunately falls a bit short of being able to assure
this.

Let's instead add a small tool runnable with `cargo xtask` to generate
sufficiently-long randomly-generated identifiers to make it easy for
people to do the right thing here for new events.